app/requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-admin-rangefilter==0.8.4
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==22.1
django-import-export==2.8.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
fake-headers==1.0.2
fake-useragent==0.1.11
flake8==4.0.1
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==4.6.11
lxml==4.9.0
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
parse==1.19.0
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyflakes==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-html==0.10.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.9
vine==1.3.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
